# backend/src/main/resources/api-docs.yml
openapi: 3.0.0
info:
  title: API Sistema Predictivo de Mantenimiento de Equipos
  version: 1.0.0
paths:
  /measurements:
    post:
      summary: Registrar una nueva medición
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Measurement'
      responses:
        '200':
          description: Medición registrada exitosamente con predicción.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasurementGraphicsDTO'
        '400':
          description: Datos de medición inválidos.
        '500':
          description: Error interno del servidor.
    get:
      summary: Obtener datos de sensores
      parameters:
        - in: query
          name: equipmentId
          required: true
          schema:
            type: string
        - in: query
          name: startDate
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: endDate
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Datos de sensores obtenidos exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Measurement'
        '404':
          description: No se encontraron mediciones para el equipo especificado.
...
components:
  schemas:
    Measurement:
      type: object
      properties:
        id:
          type: string
        equipmentId:
          type: string
        registrationDate:
          type: string
          format: date-time
        frequency:
          type: number
        current:
          type: number
        internalPressure:
          type: number
        externalPressure:
          type: number
        internalTemperature:
          type: number
        externalTemperature:
          type: number
        vibrationX:
          type: number
...
    MeasurementGraphicsDTO:
      type: object
      properties:
        id:
          type: string
        equipmentId:
          type: string
        registrationDate:
          type: string
          format: date-time
        frequency:
          type: number
        current:
          type: number
        internalPressure:
          type: number
        externalPressure:
          type: number
        internalTemperature:
          type: number
        externalTemperature:
          type: number
        vibrationX:
          type: number
        predictiveEventType:
          type: string
        probability:
          type: number
          format: double